Adjust: Ajuste nas colunas do modelo User e Movimentacao

+++ src/main/java/br/fintrack/repositories/UserRepository.java
@@ -6,5 +6,5 @@ import org.springframework.stereotype.Repository;
 import br.fintrack.models.User;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, Short> {
 }
